Remove outliers (#103)
* Update __init__.py

* Merge indicators and country profiles

* Finalize the code

* Strip subtopic

* Remove data availability + add outliers toggle + finalize data query
